Move LocationChange component under Trans Permit
