+++ ppr-ui/src/components/mhrTransfers/index.ts
@@ -6,5 +6,4 @@ export { default as TransferType } from './TransferType.vue'
 export { default as DeathCertificate } from './DeathCertificate.vue'
 export { default as SupportingDocuments } from './SupportingDocuments.vue'
 export { default as HomeOwnersGroupError } from './HomeOwnersGroupError.vue'
-export { default as LocationChange } from './LocationChange.vue'
 export { default as TaxCertificate } from './TaxCertificate.vue'
