+++ ppr-ui/src/components/mhrTransportPermit/index.ts
@@ -1,0 +1,1 @@
+export { default as LocationChange } from './LocationChange.vue'
